refactor: remove v1 admission api, refactor service method name
